requirements.txt
azure-identity==1.12.0
azure-storage-blob==12.15.0
beautifulsoup4==4.11.2
python-dotenv==1.0.0
requests==2.28.2